feat(plugin): introduce interactive UI mode for plugin features
This update introduces a new interaction type for plugin features: "UI Mode". When a user executes a plugin feature configured for UI mode in the CoreBox, it now loads and displays a custom user interface (an HTML file) directly within the CoreBox window.

- Implemented using Electron's `BrowserView` to embed the plugin UI, providing a seamless and rich interactive experience.
- Added a new `interaction` field to features in `manifest.json` to declare the UI mode and its view path.
- Extended the CoreBox manager and window manager in the main process to support `BrowserView` attachment, detachment, and lifecycle management.

This commit also includes a major refactoring of the main process plugin system.

- The monolithic `plugin-core.ts` file has been broken down into smaller, single-responsibility modules (e.g., `plugin-feature.ts`) to improve maintainability and scalability.
- Fixed an issue where the `pollingService` was not stopped on application quit.

BREAKING CHANGE: The plugin core module in the main process has been refactored. The large `plugin-core.ts` file has been removed, and its functionality is now split across multiple new files under `apps/core-app/src/main/plugins/`. Imports relying on the old file structure will be broken.
