+++ apps/core-app/src/main/plugins/index.ts
@@ -1,0 +1,4 @@
+export * from './plugin-icon'
+export * from './plugin-feature'
+export * from './plugin'
+export * from './plugin-manager'
