+++ apps/core-app/src/main/plugins/plugin-feature.ts
@@ -1,57 +1,90 @@
-import * as vm from 'vm'
-import * as fse from 'fs-extra'
-import path from 'path'
-import type { ITouchPlugin } from '@talex-touch/utils/plugin'
+import {
+  IFeatureCommand,
+  IFeatureInteraction,
+  IPluginDev,
+  IPluginFeature,
+  IPlatform,
+  IPluginIcon
+} from '@talex-touch/utils/plugin'
+import { TuffItemBuilder } from '@talex-touch/utils/core-box'
+import { PluginIcon } from './plugin-icon'
+import vm from 'vm'
+import { ITouchPlugin } from '@talex-touch/utils/plugin'
 
-/**
- * Creates a VM context for a plugin and injects a wrapped console
- * @param pluginName Plugin's unique name
- * @param featurePath Path to plugin's entry file
- * @returns The sandboxed plugin lifecycle object
- */
-export function loadPluginFeatureContext(
-  plugin: ITouchPlugin,
-  featurePath: string,
-  addonContext: any
-): any {
-  const pluginLogger = plugin.logger
-
-  const sandboxConsole = {
-    log: (...args: any[]) => {
-      console.log(...args)
-      pluginLogger.info(...args)
-    },
-    info: (...args: any[]) => {
-      console.info(...args)
-      pluginLogger.info(...args)
-    },
-    warn: (...args: any[]) => {
-      console.warn(...args)
-      pluginLogger.warn(...args)
-    },
-    error: (...args: any[]) => {
-      console.error(...args)
-      pluginLogger.error(...args)
-    },
-    debug: (...args: any[]) => {
-      console.debug(...args)
-      pluginLogger.debug?.(...args) ?? pluginLogger.info(...args)
+export const createBuilderWithPluginContext = (pluginName: string): typeof TuffItemBuilder => {
+  return class TuffItemBuilderWithPlugin extends TuffItemBuilder {
+    constructor(id: string) {
+      // Call the parent constructor and immediately set the pluginName in meta.
+      super(id, 'plugin', 'plugin-features')
+      this.setMeta({ pluginName })
     }
   }
+}
 
-  const context = vm.createContext({
-    ...addonContext,
-    console: sandboxConsole,
+import fse from 'fs-extra'
+
+export function loadPluginFeatureContextFromContent(
+  plugin: ITouchPlugin,
+  scriptContent: string,
+  context: any
+): any {
+  const sandbox = {
+    exports: {},
+    module: { exports: {} },
     require,
-    module,
-    exports,
-    __dirname: path.dirname(featurePath),
-    __filename: featurePath
-  })
+    __dirname: plugin.pluginPath,
+    __filename: 'index.js',
+    ...context
+  }
 
-  const featureCode = fse.readFileSync(featurePath, 'utf-8')
-  const script = new vm.Script(featureCode)
+  vm.createContext(sandbox)
+  vm.runInContext(scriptContent, sandbox)
 
-  const result = script.runInContext(context)
-  return result
+  return sandbox.module.exports
+}
+
+export function loadPluginFeatureContext(
+  plugin: ITouchPlugin,
+  featureIndex: string,
+  context: any
+): any {
+  const scriptContent = fse.readFileSync(featureIndex, 'utf-8')
+  return loadPluginFeatureContextFromContent(plugin, scriptContent, context)
+}
+
+export class PluginFeature implements IPluginFeature {
+  id: string
+  name: string
+  desc: string
+  icon: IPluginIcon
+  push: boolean
+  platform: IPlatform
+  commands: IFeatureCommand[]
+  interaction?: IFeatureInteraction
+  dev: IPluginDev
+
+  constructor(pluginPath: string, options: IPluginFeature, dev: IPluginDev) {
+    this.id = options.id
+    this.name = options.name
+    this.desc = options.desc
+    this.icon = new PluginIcon(pluginPath, options.icon.type, options.icon.value, dev)
+    this.push = options.push
+    this.platform = options.platform
+    this.commands = [...options.commands]
+    this.interaction = options.interaction
+    this.dev = dev
+  }
+
+  toJSONObject(): object {
+    return {
+      id: this.id,
+      name: this.name,
+      desc: this.desc,
+      icon: this.icon,
+      push: this.push,
+      platform: this.platform,
+      commands: this.commands,
+      interaction: this.interaction
+    }
+  }
 }
