+++ apps/docs/docs/plugins/config/manifest.md
@@ -61,7 +61,7 @@ package.json
     "plugin": {
       "dev": {
       "enable": true,
-      "address": "http://127.0.0.1:5174/"
+      "address": "http://127.0.0.1:3333/"
       }
     },
     "build": {}
