+++ plugins/ui-mode-tester/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+  // onFeatureTriggered is no longer needed for 'view' mode features.
+  // The CoreBoxManager handles it automatically.
+  // You can still use it for 'command' mode features.
+};
