+++ plugins/ui-mode-tester/manifest.json
@@ -1,0 +1,29 @@
+{
+  "name": "ui-mode-tester",
+  "version": "0.0.1",
+  "description": "A simple plugin to test the CoreBox UI mode.",
+  "icon": {
+    "type": "text",
+    "value": "T"
+  },
+  "features": [
+    {
+      "id": "multi-source-translate",
+      "name": "多源翻译",
+      "desc": "Enter a view for multi-source translation",
+      "icon": {
+        "type": "text",
+        "value": "文"
+      },
+      "platform": {},
+      "commands": [
+        "多源翻译",
+        "translate"
+      ],
+      "interaction": {
+        "type": "view",
+        "path": "test.html"
+      }
+    }
+  ]
+}
